+++ lib/features/survey/data.dart
@@ -82,9 +82,6 @@ class SurveyData {
 
       final resources = await getResourcesInContainer(dirUrl);
 
-      debugPrint('SubDirs: |${resources.subDirs.join('|')}|');
-      debugPrint('Files  : |${resources.files.join('|')}|');
-
       // Process each file in the directory.
 
       for (var fileName in resources.files) {